+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 **Unreleased**
 --------------
 
+- Improve moshi-sealed KSP error messages.
 - Fix fallback adapter support in IR code gen not recognizing Moshi parameters to primary constructors.
 
 0.27.0